% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{ancestors_graph}
\alias{ancestors_graph}
\title{Return an igraph of a vertex and all its ancestors to the root}
\usage{
ancestors_graph(g, v)
}
\arguments{
\item{g}{The graph from which to get ancestors of vertex v}

\item{v}{The vertex of g from which to get 'upstream' vertices}
}
\value{
A subgraph of g with v and its ancestors
}
\description{
Return an igraph of a vertex and all its ancestors to the root
}
\examples{
ancestors_graph(NVC_graph, "856963")
}
\seealso{
\link{children_graph}
}

