% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{children_graph}
\alias{children_graph}
\title{Return an igraph of a vertex and all its children}
\usage{
children_graph(g, v)
}
\arguments{
\item{g}{The graph from which to get children of vertex v}

\item{v}{The vertex of g from which to get 'downstream' vertices}
}
\value{
A subgraph of g with v and its children
}
\description{
More detailed description of the function
}
\examples{
children_graph(NVC_graph, "856963")
}
\seealso{
\link{ancestor_graph}
}

