% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d_classif_confidence}
\alias{d_classif_confidence}
\title{Relates confidence IDs to their description}
\format{A tibble with 5 rows and 4 columns
\describe{
\item{\code{D_CLASSIF_CONFIDENCE_ID}}{Code for classification confidence}
\item{\code{CLASSIF_CONFIDENCE_DESC}}{Description of classification
            confidence}
\item{\code{DISPLAY_ORDER}}{A factor corresponding to the level of the
            NVC hierarchy}
\item{\code{STATUS}}{Description of Accepted/Proposed}
}}
\source{
https://www1.usgs.gov/csas/nvcs_data/USNVCFullDownload.zip
}
\usage{
d_classif_confidence
}
\description{
Relates confidence IDs to their description
}
\keyword{datasets}

