% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d_classification_level}
\alias{d_classification_level}
\title{Relationships of classification levels}
\format{A data.frame with 16 rows and 4 columns
\describe{
\item{\code{D_CLASSIF_LEVEL_ID}}{Code for classification level}
\item{\code{CLASSIFICATION_LEVEL_NAME}}{Name of the classification level}
\item{\code{DISPLAY_ORDER}}{A factor corresponding to the level of the
            NVC hierarchy}
\item{\code{parent_level_id}}{ID of parent to classification level}
}}
\source{
https://www1.usgs.gov/csas/nvcs_data/USNVCFullDownload.zip
}
\usage{
d_classification_level
}
\description{
Relationships of classification levels
}
\keyword{datasets}

