% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d_curr_presence_absence}
\alias{d_curr_presence_absence}
\title{Relationships of presence/absence of NVC categories}
\format{A data.frame with 3 rows and 3 columns
\describe{
\item{\code{D_CURR_PRESENCE_ABSENCE_ID}}{Code for presence/absence}
\item{\code{CURR_PRESENCE_ABSENCE_DESC}}{Description of presence/absence code}
\item{\code{CURR_PRESENCE_ABSENCE_CD}}{Unknown}
}}
\source{
https://www1.usgs.gov/csas/nvcs_data/USNVCFullDownload.zip
}
\usage{
d_curr_presence_absence
}
\description{
Relationships of presence/absence of NVC categories
}
\keyword{datasets}

