% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d_dist_confidence}
\alias{d_dist_confidence}
\title{Relationships of distribution codes for NVC categories}
\format{A data.frame with 7 rows and 3 columns
\describe{
\item{\code{D_DIST_CONFIDENCE_ID}}{Numeric code for distribution confidence}
\item{\code{DIST_CONFIDENCE_CD}}{Character code for distribution confidence}
\item{\code{DIST_CONFIDENCE_DESC}}{Description of distribution confidence}
}}
\source{
https://www1.usgs.gov/csas/nvcs_data/USNVCFullDownload.zip
}
\usage{
d_dist_confidence
}
\description{
Relationships of distribution codes for NVC categories
}
\keyword{datasets}

