% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d_occurrence_status}
\alias{d_occurrence_status}
\title{Relationships of occurrence codes for NVC categories}
\format{A data.frame with 6 rows and 3 columns
\describe{
\item{\code{D_OCCURRENCE_STATUS_ID}}{Numeric code for occurrence status}
\item{\code{OCCURRENCE_STATUS_CD}}{Character code for occurrence status}
\item{\code{OCCURRENCE_STATUS_DESC}}{Description of occurrence status}
}}
\source{
https://www1.usgs.gov/csas/nvcs_data/USNVCFullDownload.zip
}
\usage{
d_occurrence_status
}
\description{
Relationships of occurrence codes for NVC categories
}
\keyword{datasets}

