% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d_subnation}
\alias{d_subnation}
\title{Relationships of subnation codes for NVC categories}
\format{A data.frame with 77 rows and 4 columns
\describe{
\item{\code{SUBNATION_ID}}{Numeric code for subnation entity}
\item{\code{ISO_NATION_CD}}{Character ISO code for nation entity}
\item{\code{SUBNATION_CODE}}{Character code (abbrev.) for subnation entity}
\item{\code{SUBNATION_NAME}}{Text description of subnation entity}
}}
\source{
https://www1.usgs.gov/csas/nvcs_data/USNVCFullDownload.zip
}
\usage{
d_subnation
}
\description{
Relationships of subnation codes for NVC categories
}
\keyword{datasets}

