% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d_usfs_ecoregion}
\alias{d_usfs_ecoregion}
\title{Relationships of USDA Forest Service ecoregion fields}
\format{A data.frame with 771 rows and 6 columns
\describe{
\item{\code{USFS_ECOREGION_ID}}{Numeric code for USFS ecoregion}
\item{\code{PARENT_USFS_ECOREGION_ID}}{Parent ID for USFS ecoregion}
\item{\code{D_USFS_ECOREGION_LEVEL_ID}}{Numeric code of classification level}
\item{\code{USFS_ECOREGION_NAME}}{Name of USFS ecoregion}
\item{\code{USFS_ECOREGION_CLASS_CD}}{Character code for USFS class (level)}
\item{\code{USFS_ECOREGION_CONCAT_CD}}{Character code for USFS class +
            ecoregion, concatenated}
}}
\source{
https://www1.usgs.gov/csas/nvcs_data/USNVCFullDownload.zip
}
\usage{
d_usfs_ecoregion
}
\description{
Unclear whether the 2007 Forest Service ecoregions are newer or older than
the 'plain' Forest Service ecoregions. Also unclear where the values in
USFS_ECOREGION_CONCAT_CD come from.
}
\keyword{datasets}

