% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d_usfs_ecoregion2007}
\alias{d_usfs_ecoregion2007}
\title{Relationships of 2007 USDA Forest Service ecoregion fields}
\format{A data.frame with 1506 rows and 5 columns
\describe{
\item{\code{USFS_ECOREGION_ID}}{Numeric code for 2007 USFS ecoregion}
\item{\code{PARENT_USFS_ECOREGION_ID}}{Parent ID for 2007 USFS ecoregion}
\item{\code{D_USFS_ECOREGION_LEVEL_ID}}{Numeric code of classification level}
\item{\code{USFS_ECOREGION_NAME}}{Name of 2007 USFS ecoregion}
\item{\code{USFS_ECOREGION_CONCAT_CD}}{Character code for USFS level +
            ecoregion, concatenated}
}}
\source{
https://www1.usgs.gov/csas/nvcs_data/USNVCFullDownload.zip
}
\usage{
d_usfs_ecoregion2007
}
\description{
Unclear whether the 2007 Forest Service ecoregions are newer or older than
the 'plain' Forest Service ecoregions. Also unclear where the values in
USFS_ECOREGION_CONCAT_CD come from.
}
\keyword{datasets}

