% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d_usfs_ecoregion_level}
\alias{d_usfs_ecoregion_level}
\title{Relationships of USDA Forest Service ecoregion levels}
\format{A data.frame with 8 rows and 3 columns
\describe{
\item{\code{D_USFS_ECOREGION_LEVEL_ID}}{Numeric code for USFS ecoregion level}
\item{\code{USFS_ECOREGION_LEVEL_DESC}}{Description of the ecoregion level}
\item{\code{D_USFS_ECOREGION_LEVEL_ID}}{Numeric code of parent ecoregion}
}}
\source{
https://www1.usgs.gov/csas/nvcs_data/USNVCFullDownload.zip
}
\usage{
d_usfs_ecoregion_level
}
\description{
Unclear whether this is for Forest Service ecoregions 'plain' or 2007, but
presumably for the 'plain' version
}
\keyword{datasets}

