% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ancestor_text}
\alias{get_ancestor_text}
\title{Return the concatenated text fields for ancestors and a vertex}
\usage{
get_ancestor_text(g, v, ...)
}
\arguments{
\item{g}{The graph to be queried}

\item{v}{The NVC category for which the 'ancestors+' description will be
created}

\item{...}{One or more text variables to concatenate together}
}
\value{
A string of all ancestral typeConcepts of v, concatenated with ' | '
}
\description{
Return the concatenated text fields for ancestors and a vertex
}
\examples{
get_ancestor_text(NVC_graph, "860275", "typeDesc", "Environment")
}

