% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ancestor_textDesc}
\alias{get_ancestor_textDesc}
\title{Return the concatenated text descriptions for ancestors and a vertex}
\usage{
get_ancestor_textDesc(g, v)
}
\arguments{
\item{g}{The graph to be queried}

\item{v}{The NVC category for which the 'ancestors+' description will be
created}
}
\value{
A string of all relevant descriptive text, concatenated with ' | '
}
\description{
Return the concatenated text descriptions for ancestors and a vertex
}
\examples{
get_ancestor_textDesc(NVC_graph, "860275")
}

