% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ancestor_typeConcept}
\alias{get_ancestor_typeConcept}
\title{Return the concatenated typeDescription for ancestors and a vertex}
\usage{
get_ancestor_typeConcept(g, v)
}
\arguments{
\item{g}{The graph to be queried}

\item{v}{The NVC category for which the 'ancestors+' description will be
created}
}
\value{
A string of all ancestral typeConcepts of v, concatenated with ' | '
}
\description{
Return the concatenated typeDescription for ancestors and a vertex
}
\examples{
get_ancestor_typeConcept(NVC_graph, "860275")
}

