% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{get_ancestors}
\alias{get_ancestors}
\title{Return a vector of all 'upstream' vertices in a graph from a given vertex}
\usage{
get_ancestors(g, v)
}
\arguments{
\item{g}{The graph from which to get the children}

\item{v}{The `name` of the vertex to get ancestors for}
}
\value{
A vector of the `name`s of ancestor vertices from v to root Class
}
\description{
Return a vector of all 'upstream' vertices in a graph from a given vertex
}
\examples{
\dontrun{
get_ancestors(NVC_graph, "856963")
}
}
\seealso{
\link{get_ancestors}
}

