% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{get_children}
\alias{get_children}
\title{Return a vector of all 'downstream' vertices in a graph from a given vertex}
\usage{
get_children(g, v)
}
\arguments{
\item{g}{The graph from which to get the children}

\item{v}{The `name` of the vertex to get children for}
}
\value{
A vector of the `name`s of children vertices from v
}
\description{
Return a vector of all 'downstream' vertices in a graph from a given vertex
}
\examples{
\dontrun{
get_children(NVC_graph, "856963")
}
}
\seealso{
\link{get_ancestors}
}

