% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_full_textDesc}
\alias{get_full_textDesc}
\title{Return the concatenated text descriptions for ancestors and children of a vertex}
\usage{
get_full_textDesc(g, v)
}
\arguments{
\item{g}{The graph to be queried}

\item{v}{The NVC category for which the 'full' description will be created}
}
\value{
A string of all relevant descriptive text, concatenated with ' | '
}
\description{
Return the concatenated text descriptions for ancestors and children of a vertex
}
\examples{
get_full_textDesc(NVC_graph, "860275")
}

