% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_hier_desc}
\alias{get_hier_desc}
\title{Get description data for a level of the NVC hierarchy}
\usage{
get_hier_desc(level)
}
\arguments{
\item{level}{The hierarchical level of the NVC to extract}
}
\value{
A data.frame with global ID, colloquial name, and the type concept
}
\description{
For natural assemblages, level is one of: Class, Subclass, Formation,
Division, Macrogroup, Group, Alliance, Association. For human-created
features, one of: Cultural Class, Cultural Subclass, Cultural Formation,
Cultural Subformation, Cultural Type, Cultural Group, Cultural Subgroup
}
\examples{
get_hier_desc("Class")
}

