% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{get_root_path}
\alias{get_root_path}
\title{Get the root Class for any given leaf vertex}
\usage{
get_root_path(g, v)
}
\arguments{
\item{g}{The graph to be queried}

\item{v}{A character string of the identifier node to trace back}
}
\value{
A vector of the path from the leaf node to its Class root
}
\description{
More detailed description of the function
}
\examples{
\dontrun{
get_root_path(NVC_graph, "860275")
}
}

