% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{get_top_class}
\alias{get_top_class}
\title{Return the NVC Class to which a given vertex belongs}
\usage{
get_top_class(g, v)
}
\arguments{
\item{g}{The graph to query}

\item{v}{The vertex ID ('name' of attr) to trace back to Class}
}
\value{
The colloquialName of the Class to which vertex v belongs
}
\description{
Return the NVC Class to which a given vertex belongs
}
\examples{
get_top_class(NVC_graph, "860285")
}
\seealso{
if any see alsos
}

