% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph_to_depth}
\alias{graph_to_depth}
\title{Return the NVC graph out from the "center" at natural and cultural domains}
\usage{
graph_to_depth(g, depth)
}
\arguments{
\item{g}{The graph from which to get the subgraph}

\item{depth}{The distance from the center of g to extend the subgraph}
}
\value{
An igraph subgraph of g, to distance `depth`
}
\description{
Useful, perhaps, for plotting a small version of the full NVC_graph.
}

