% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{make_igraph}
\alias{make_igraph}
\title{Make an igraph representation of the NVC}
\usage{
make_igraph(df, directed = TRUE)
}
\arguments{
\item{df}{The NVC data.frame (or subset thereof) to convert to an igraph graph}

\item{directed}{Whether to return a directed graph (default = TRUE)}
}
\value{
An \link{igraph} graph
}
\description{
Converting the tables from NVC to graph structures opens up many possibilities
for discovering novel patterns and extracting data. This uses a loop over the
parent category information to create the from-to links.
}
\examples{
\dontrun{
make_igraph(df)
}
}

