% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{path_between_cats}
\alias{path_between_cats}
\title{Find the path between nodes in the NVC graph}
\usage{
path_between_cats(g, node1, node2)
}
\arguments{
\item{g}{The graph to traverse}

\item{node1}{The starting node in the path}

\item{node2}{The end node in the path}
}
\value{
A list of the node (categories) on the path from node1 to node2
}
\description{
More detailed description of the function
}
\examples{
path_between_cats(NVC_graph, "685697", "686373")
}

