% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prep_for_visNetwork}
\alias{prep_for_visNetwork}
\title{Create data.frames needed for visNetwork visualization}
\usage{
prep_for_visNetwork(g)
}
\arguments{
\item{g}{The graph to be visualized}
}
\value{
A list of two dataframes, nodes and edges
}
\description{
\code{visNetwork} has a helper function for igraph structures,
\code{visIgraph}, but setting certain attributes (e.g., labels) doesn't seem
to work well. This is a work-around.
}
\examples{
prep_for_visNetwork(graph_to_depth(NVC_graph, 4))
}
\seealso{
if any see alsos
}

