% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prep_text}
\alias{prep_text}
\title{Clean corpus text for NLP}
\usage{
prep_text(corp)
}
\arguments{
\item{corp}{A corpus from \link[tm]{tm}}
}
\value{
A corpus with puctuation, numbers, stopwords, and whitespace removed
}
\description{
Clean corpus text for NLP
}
\examples{
\dontrun{
prep_text(a_corpus)
}
}
\seealso{
if any see alsos
}

