% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Reference}
\alias{Reference}
\title{References cited in the NVC}
\format{A data.frame with 6659 rows and 3 columns
\describe{
\item{\code{REFERENCE_ID}}{Numeric code for the reference}
\item{\code{SHORTCITATION}}{Author Year representation for the reference}
\item{\code{FULLCITATION}}{Full text citation; unclear if this follows an
            established citation style.}
}}
\source{
https://www1.usgs.gov/csas/nvcs_data/USNVCFullDownload.zip
}
\usage{
Reference
}
\description{
References cited in the NVC
}
\keyword{datasets}

