% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Unit}
\alias{Unit}
\title{NVC Unit table as a data.frame}
\format{A tibble with 8192 rows and 14 columns.
\describe{
\item{\code{ELEMENT_GLOBAL_ID}}{The unique identifier of the NVC category}
\item{\code{PARENT_ID}}{The identifier of the parent of the NVC category}
\item{\code{D_CLASSIFICATION_LEVEL_ID}}{A factor corresponding to the level
            of the hierarchy}
\item{\code{elementUID}}{ELEMENT_GLOBAL_ID prefixed with '2.', presumably to
            identify it as deriving from version 2 of the standard}
\item{\code{classificationCode}}{An alpha-numeric sequence indicating the
            placement of the category in the NVC hierarchy}
\item{\code{databaseCode}}{An identifier for creating an SQL database}
\item{\code{Status}}{Whether the category has been accepted into the NVC}
\item{\code{colloquialName}}{A 'common' name for the category, if available}
\item{\code{scientificName}}{A name for the category, which may include
            scientific names of characteristic taxa}
\item{\code{formattedScientificName}}{'scientificName' with HTML formatting}
\item{\code{translatedName}}{'scientificName' with common name substitution}
\item{\code{hierarchyLevel}}{Name of the level of the hierarchy}
\item{\code{unitSort}}{An alpha-numeric code indicating the placement of the
            category in the NVC; different from 'classificationCode'}
\item{\code{USstatus}}{A code ???}
}}
\source{
https://www1.usgs.gov/csas/nvcs_data/USNVCFullDownload.zip
}
\usage{
Unit
}
\description{
The U.S. National Vegetation Classification (NVC) is provided as a series of
tables at \url{http://usnvc.org/explore-classification/}.
}
\keyword{datasets}

