% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unit_data}
\alias{unit_data}
\title{Core NVC data (Unit & UnitDescription) as a data.frame}
\format{A data.frame with 8192 rows and 45 columns.
\describe{
\item{\code{ELEMENT_GLOBAL_ID}}{The unique identifier of the NVC category}
\item{\code{PARENT_ID}}{The identifier of the parent of the NVC category}
\item{\code{D_CLASSIFICATION_LEVEL_ID}}{A factor corresponding to the level
            of the hierarchy}
\item{\code{elementUID}}{ELEMENT_GLOBAL_ID prefixed with '2.', presumably to
            identify it as deriving from version 2 of the standard}
\item{\code{classificationCode}}{An alpha-numeric sequence indicating the
            placement of the category in the NVC hierarchy}
\item{\code{databaseCode}}{An identifier for creating an SQL database}
\item{\code{Status}}{Whether the category has been accepted into the NVC}
\item{\code{colloquialName}}{A 'common' name for the category, if available}
\item{\code{scientificName}}{A name for the category, which may include
            scientific names of characteristic taxa}
\item{\code{formattedScientificName}}{'scientificName' with HTML formatting}
\item{\code{translatedName}}{'scientificName' with common name substitution}
\item{\code{hierarchyLevel}}{Name of the level of the hierarchy}
\item{\code{unitSort}}{An alpha-numeric code indicating the placement of the
            category in the NVC; different from 'classificationCode'}
\item{\code{USstatus}}{A code ???}
\item{\code{parentKey}}{The 'databaseCode' of the parent of a category}
\item{\code{parentName}}{Another representation of the parent of a category}
\item{\code{typeConceptSentence}}{A concise sentence describing the category}
\item{\code{typeConcept}}{Text description of the conceptual model of the
            category}
\item{\code{classificationComments}}{Comments about the classification}
\item{\code{similarNVCtypesComments}}{Comments referencing similar categories}
\item{\code{diagnosticCharacteristics}}{Text description (with HTML tags) of
            the characteristics that differentiate the category}
\item{\code{Rationale}}{Unknown; all NA}
\item{\code{Physiognomy}}{Text description (with HTML tags) of the
            physiognomic characteristics (structure) of the category}
\item{\code{Floristics}}{Text description (with HTML tags) of the
            floral characteristics (species/groups) of the category}
\item{\code{Dynamics}}{Text description (with HTML tags) of the
            physical/ecological dynamics of the category}
\item{\code{Environment}}{Text description (with HTML tags) of the
            environmental characteristics (e.g., climate) of the category}
\item{\code{Analysis}}{Unknown; all NA}
\item{\code{Range}}{Text description of the category's geographical spread
            and/or relationship to other classifications}
\item{\code{spatialPattern}}{Categorical designation of how the NVC category
            is typically arranged, e.g., large patch vs. linear}
\item{\code{Nations}}{Comma-separated string of two-letter country codes
            indicating nations where the category is found}
\item{\code{Subnations}}{Comma-separated string of two-letter state and
            province codes indicating where the category is found}
\item{\code{federalLands}}{Semicolon-separated string of agency and site
            where the category is found}
\item{\code{tncEcoregions}}{Comma-separated string relating the category to
            one or more TNC ecoregion}
\item{\code{usfsEcoregions}}{Comma-separated string relating the category to
            one or more USDA Forest Service ecoregion}
\item{\code{omernikEcoregions}}{Unknown; all NA}
\item{\code{classif_confidence_id}}{A code describing Low:High confidence of
            the category classification; see \link{d_classif_confidence}}
\item{\code{confidenceComments}}{Text describing how/why the confidence level
            was assigned; mostly empty}
\item{\code{Synonymy}}{Text (with HTML tags) describing synonyms or approx.
            synonyms of the category to other classifications}
\item{\code{primaryConceptSource}}{Reference of the authority of the concept}
\item{\code{descriptionAuthor}}{The person/entity who authored the
            'description' (perhaps the typeConcept?)}
\item{\code{Acknowledgements}}{Text giving acknowledgement}
\item{\code{versionDate}}{Text date stamp, format = mm/dd/yyy}
\item{\code{CLASSIFICATION_LEVEL_NAME}}{Name of the level of the hierarchy;
            see also hierarchyLevel}
\item{\code{DISPLAY_ORDER}}{Numeric indicator of the category's placement in
            the hierarchy}
\item{\code{parent_level_id}}{Numeric indicator of the category's parent
            category}
}}
\source{
https://www1.usgs.gov/csas/nvcs_data/USNVCFullDownload.zip
}
\usage{
unit_data
}
\description{
The U.S. National Vegetation Classification (NVC) is provided as a series of
tables at \url{http://usnvc.org/explore-classification/}. This table is
from a \code{JOIN} of the Unit and UnitDescriptions database tables.
}
\keyword{datasets}

