% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UnitDescription}
\alias{UnitDescription}
\title{NVC UnitDescription table as a data.frame}
\format{A tibble with 8192 rows and 29 columns.
\describe{
\item{\code{ELEMENT_GLOBAL_ID}}{The unique identifier of the NVC category}
\item{\code{parentKey}}{The 'databaseCode' of the parent of a category}
\item{\code{parentName}}{Another representation of the parent of a category}
\item{\code{typeConceptSentence}}{A concise sentence describing the category}
\item{\code{typeConcept}}{Text description of the conceptual model of the
            category}
\item{\code{classificationComments}}{Comments about the classification}
\item{\code{similarNVCtypesComments}}{Comments referencing similar categories}
\item{\code{diagnosticCharacteristics}}{Text description (with HTML tags) of
            the characteristics that differentiate the category}
\item{\code{Rationale}}{Unknown; all NA}
\item{\code{Physiognomy}}{Text description (with HTML tags) of the
            physiognomic characteristics (structure) of the category}
\item{\code{Floristics}}{Text description (with HTML tags) of the
            floral characteristics (species/groups) of the category}
\item{\code{Dynamics}}{Text description (with HTML tags) of the
            physical/ecological dynamics of the category}
\item{\code{Environment}}{Text description (with HTML tags) of the
            environmental characteristics (e.g., climate) of the category}
\item{\code{Analysis}}{Unknown; all NA}
\item{\code{Range}}{Text description of the category's geographical spread
            and/or relationship to other classifications}
\item{\code{spatialPattern}}{Categorical designation of how the NVC category
            is typically arranged, e.g., large patch vs. linear}
\item{\code{Nations}}{Comma-separated string of two-letter country codes
            indicating nations where the category is found}
\item{\code{Subnations}}{Comma-separated string of two-letter state and
            province codes indicating where the category is found}
\item{\code{federalLands}}{Semicolon-separated string of agency and site
            where the category is found}
\item{\code{tncEcoregions}}{Comma-separated string relating the category to
            one or more TNC ecoregion}
\item{\code{usfsEcoregions}}{Comma-separated string relating the category to
            one or more USDA Forest Service ecoregion}
\item{\code{omernikEcoregions}}{Unknown; all NA}
\item{\code{classif_confidence_id}}{A code describing Low:High confidence of
            the category classification; see \link{d_classif_confidence}}
\item{\code{confidenceComments}}{Text describing how/why the confidence level
            was assigned; mostly empty}
\item{\code{Synonymy}}{Text (with HTML tags) describing synonyms or approx.
            synonyms of the category to other classifications}
\item{\code{primaryConceptSource}}{Reference of the authority of the concept}
\item{\code{descriptionAuthor}}{The person/entity who authored the
            'description' (perhaps the typeConcept?)}
\item{\code{Acknowledgements}}{Text giving acknowledgement}
\item{\code{versionDate}}{Text date stamp, format = mm/dd/yyy}
}}
\source{
https://www1.usgs.gov/csas/nvcs_data/USNVCFullDownload.zip
}
\usage{
UnitDescription
}
\description{
The U.S. National Vegetation Classification (NVC) is provided as a series of
tables at \url{http://usnvc.org/explore-classification/}.
}
\keyword{datasets}

