% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UnitXEcoregionUsfs}
\alias{UnitXEcoregionUsfs}
\title{Relationships of USDA Forest Service ecoregions to NVC categories}
\format{A data.frame with 45914 rows and 3 columns
\describe{
\item{\code{ELEMENT_GLOBAL_ID}}{The unique identifier of the NVC category}
\item{\code{USFS_ECOREGION_ID}}{Numeric code for USFS ecoregion}
\item{\code{D_OCCURRENCE_STATUS_ID}}{Numeric code for occurrence status}
}}
\source{
https://www1.usgs.gov/csas/nvcs_data/USNVCFullDownload.zip
}
\usage{
UnitXEcoregionUsfs
}
\description{
Appears to be an expansion of the \code{usfsEcoregions} field in
\link{UnitDescription} with information on occurrence added in.
}
\keyword{datasets}

