% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UnitXEcoregionUsfs2007}
\alias{UnitXEcoregionUsfs2007}
\title{Relationships of 2007 USDA Forest Service ecoregions to NVC categories}
\format{A data.frame with 10484 rows and 3 columns
\describe{
\item{\code{ELEMENT_GLOBAL_ID}}{The unique identifier of the NVC category}
\item{\code{USFS_ECOREGION_2007_ID}}{Numeric code for 2007 USFS ecoregion}
\item{\code{D_OCCURRENCE_STATUS_ID}}{Numeric code for occurrence status}
}}
\source{
https://www1.usgs.gov/csas/nvcs_data/USNVCFullDownload.zip
}
\usage{
UnitXEcoregionUsfs2007
}
\description{
Unclear of the provenance of this data; may be a multi-way \code{JOIN}.
}
\keyword{datasets}

