% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UnitXReference}
\alias{UnitXReference}
\title{Relationships of references to NVC categories}
\format{A data.frame with 77150 rows and 2 columns
\describe{
\item{\code{ELEMENT_GLOBAL_ID}}{The unique identifier of the NVC category}
\item{\code{REFERENCE_ID}}{Numeric code for reference}
}}
\source{
https://www1.usgs.gov/csas/nvcs_data/USNVCFullDownload.zip
}
\usage{
UnitXReference
}
\description{
Relationships of references to NVC categories
}
\keyword{datasets}

