% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UnitXSimilarUnit}
\alias{UnitXSimilarUnit}
\title{Relationships among NVC categories (similarities)}
\format{A data.frame with 11739 rows and 2 columns
\describe{
\item{\code{ELEMENT_GLOBAL_ID}}{The unique identifier of the NVC category}
\item{\code{SIMGLOBAL_ID}}{The unique identifier of the similar NVC category}
}}
\source{
https://www1.usgs.gov/csas/nvcs_data/USNVCFullDownload.zip
}
\usage{
UnitXSimilarUnit
}
\description{
Unclear how this table was derived.
}
\keyword{datasets}

