% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UnitXSubnation}
\alias{UnitXSubnation}
\title{Relationships of subnation IDs to NVC categories}
\format{A data.frame with 32934 rows and 4 columns
\describe{
\item{\code{ELEMENT_GLOBAL_ID}}{The unique identifier of the NVC category}
\item{\code{SUBNATION_ID}}{The unique identifier of the subnation entity}
\item{\code{D_CURR_PRESENCE_ABSENCE_ID}}{Code for presence/absence}
\item{\code{D_DIST_CONFIDENCE_ID}}{Numeric code for distribution confidence}
}}
\source{
https://www1.usgs.gov/csas/nvcs_data/USNVCFullDownload.zip
}
\usage{
UnitXSubnation
}
\description{
Relationships of subnation IDs to NVC categories
}
\seealso{
\link{d_dist_confidence} \link{d_curr_presence_absence}
}
\keyword{datasets}

