% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/USNVC.R
\docType{package}
\name{USNVC}
\alias{USNVC}
\alias{USNVC-package}
\title{USNVC: A package for the U.S. National Vegetation Classification (NVC).}
\description{
The USNVC package provides simple access to the data in the U.S. National
Vegetation Classification as tables and as a graph structure, as well as
helper functions to access the data.
}
\section{USNVC overview}{

The \href{http://usnvc}{U.S. National Vegetation Classification (NVC)} is a
standardized, hierarchical classification system for vegetation in the United
States. (It is aptly named.) The \href{https://www.fgdc.gov/standards/projects/FGDC-standards-projects/vegetation/NVCS_V2_FINAL_2008-02.pdf}{NVC standards} were developed by a partnership of agencies (e.g., USDA
Forest Service and US Geological Survey), professional organizations (e.g.,
the Ecological Society of America), and non-profits (e.g., NatureServe). The
full NVC database, which can be downloaded from
\url{http://www1.usgs.gov/csas/nvcs_data/USNVCFullDownload.zip}, contains 18
tables with information on 8,192 vegetation categories, ranging in scope from
seven very broad Classes to >4,000 highly refined Associations.
}

