% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_hier_desc}
\alias{get_hier_desc}
\title{Get description data for a level of the NVC hierarchy}
\usage{
get_hier_desc(level)
}
\arguments{
\item{level}{The hierarchical level to extract}
}
\value{
A data.frame with global ID, colloquial name, and the type concept
}
\description{
More detailed description of the function
}
\examples{
get_hier_desc("Class")
}
\seealso{
if any see alsos
}

