% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{get_root_path}
\alias{get_root_path}
\title{Get the root Class for any given leaf vertex}
\usage{
get_root_path(g, leaf)
}
\arguments{
\item{g}{The graph to be queried}

\item{leaf}{A character string of the name of the leaf node to trace back}
}
\value{
A vector of the path from the leaf node to its Class root
}
\description{
More detailed description of the function
}
\examples{
\dontrun{
get_root_path("Temperate Interstitial Shore")
}
}

